timer.c: fix inline/extern for DoTimer() and ADD_OUT()
if I understand correctly, gcc without optimization makes wrong
decisions about compiling inline functions. According to the
spec, if inline function declared without static keyword, the
compiler should make and instance of it, but can inline it in the
the file where it is declared. When we see bunch of errors of
underfined symbols meaning that the compiler put calls to the
function but did not create an instance.

Make ADD_OUT() static inline since it is used only in the file
and DoTimer() -- normal exported function since it is used
outside as well.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>